+++ poetry/masonry/publishing/uploader.py
@@ -2,6 +2,8 @@ import hashlib
 import io
 import re
 
+from typing import List
+
 import requests
 
 from requests import adapters
@@ -51,6 +53,23 @@ class Uploader:
         )
 
         return adapters.HTTPAdapter(max_retries=retry)
+
+    @property
+    def files(self):  # type: () -> List[str]
+        dist = self._poetry.file.parent / 'dist'
+        version = normalize_version(self._package.version.text)
+
+        wheels = list(dist.glob(
+            '{}-{}-*.whl'.format(
+                re.sub("[^\w\d.]+", "_", self._package.pretty_name, flags=re.UNICODE),
+                re.sub("[^\w\d.]+", "_", version, flags=re.UNICODE),
+            )
+        ))
+        tars = list(dist.glob(
+            '{}-{}.tar.gz'.format(self._package.pretty_name, version)
+        ))
+
+        return sorted(wheels + tars)
 
     def auth(self, username, password):
         self._username = username
@@ -178,28 +197,7 @@ class Uploader:
                 raise
 
     def _do_upload(self, session, url):
-        dist = self._poetry.file.parent / 'dist'
-        version = normalize_version(self._package.version.text)
-        packages = dist.glob(
-            '{}-{}*'.format(self._package.name, version)
-        )
-        files = (
-            i for i in packages if (
-                i.match(
-                    '{}-{}-*.whl'.format(
-                        self._package.name, version
-                    )
-                )
-                or
-                i.match(
-                    '{}-{}.tar.gz'.format(
-                        self._package.name, version
-                    )
-                )
-        )
-        )
-
-        for file in files:
+        for file in self.files:
             # TODO: Check existence
 
             resp = self._upload_file(session, url, file)