+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 
 - Improved the `show` command to make it easier to check if packages are properly installed.
 - The `script` command has been deprecated, use `run` instead.
+- The `publish` command no longer build packages by default. Use `--build` to retrieve the previous behavior.
 - Improved support for private repositories.
 - Expanded version constraints now keep the original version's precision.
 - The lock file hash no longer use the project's name and version.